% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apelake.R
\docType{data}
\name{apelake}
\alias{apelake}
\title{Ape Lake Station Dataframe}
\format{
A data frame with 5097 rows and 24 variables:
\describe{

  \item{WatYr}{year, numeric}
  \item{DateTime}{year date and time of the observation, POSIXct}
  \item{Air_Temp}{Air temperature, in degrees celcius}
  ...
}
}
\source{
\url{http://www.viu-hydromet-wx.ca/}
}
\usage{
apelake
}
\description{
An example dataset from the apelake station that contains hydrometerological
variables for testing functions.
}
\keyword{datasets}
