% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat_wind.R
\name{cat_wind}
\alias{cat_wind}
\title{Categorize wind speed from numerical to cardinal names (N,...)}
\usage{
cat_wind(df, wind_dir_num, bin_col_name, dir_res = 45)
}
\arguments{
\item{df}{dataframe object that contains at least one column with numerical wind speed directions.}

\item{wind_dir_num}{string of column name that contains the wind directions.}

\item{bin_col_name}{string of the new column name to contain the wind direction bin.}

\item{dir_res}{integer representing the degree resolution of the bins. E.g. 45 deg. is 8 bins}
}
\value{
dataframe with additional numerical column that contains binned wind speed directions
}
\description{
Categorize wind speed from numerical to cardinal names (N,...)
}
\examples{
cat_wind(apelake, 'Wind_Dir', 'test', 45) |> dplyr::select(Wind_Dir, test)
}
