% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{graph_two}
\alias{graph_two}
\title{Create Graph With Two Parameters}
\usage{
graph_two(
  data,
  x,
  y1,
  y2,
  y1_name,
  y2_name,
  col_one = "rgb(0,119,187)",
  col_two = "rgb(204,51,17)",
  margin = list(b = 0, r = 50, l = 50, t = 10)
)
}
\arguments{
\item{data}{with posixct datetime and at least one weather parameter}

\item{x}{datetime as posixct}

\item{y1}{weather variable 1}

\item{y2}{weather variable 2}

\item{y1_name}{pretty name to be used on plot}

\item{y2_name}{pretty name to be used on plot}

\item{col_one}{colour for line ex. "rgb(0,119,187)"}

\item{col_two}{colour for line ex. "rgb(0,119,187)"}

\item{margin}{standard margins for plotly plot}
}
\value{
plotly graph
}
\description{
Create Graph With Two Parameters
}
