% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temporal.R
\name{set_yr}
\alias{set_yr}
\title{set year of posix datetime using lubridate for if else statement}
\usage{
set_yr(DATE_TIME, YEAR)
}
\arguments{
\item{DATE_TIME}{datetime to set year}

\item{YEAR}{year to set}
}
\value{
posixct with updated yr
}
\description{
set year of posix datetime using lubridate for if else statement
}
