% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qaqc.R
\name{spike_clean}
\alias{spike_clean}
\title{Clean Snow Depth Data}
\usage{
spike_clean(data, datetime, snow, spike_th, roc_hi_th, roc_low_th)
}
\arguments{
\item{data}{dataframe with snow depth data}

\item{datetime}{col for datetime to fill gaps}

\item{snow}{col name for snow depth}

\item{spike_th}{threshold to be used for pos. AND neg. spikes}

\item{roc_hi_th}{rate of change threshold}

\item{roc_low_th}{rate of change threshold}
}
\value{
cleaned dataframe
}
\description{
Clean Snow Depth Data
}
