% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temporal.R
\name{wtr_yr}
\alias{wtr_yr}
\title{calculate water year by date}
\usage{
wtr_yr(dates, start_month = 10)
}
\arguments{
\item{dates}{Date 'YYYY-MM-DD'}

\item{start_month}{water year typically starts with 10 but can change if different.}
}
\value{
year
}
\description{
calculate water year by date
}
