% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{wind_rose}
\alias{wind_rose}
\title{Create Wind Rose Plot}
\usage{
wind_rose(
  data,
  wind_spd_avg,
  wind_dir_avg_avg,
  hrs = 168,
  plotTitle = "",
  dirres = 15
)
}
\arguments{
\item{data}{dataframe with datetime, wind speed and wind dir}

\item{wind_spd_avg}{name of the wind speed column}

\item{wind_dir_avg_avg}{name of the wind direction column}

\item{hrs}{the number of hours to create the plot over, 168 for weekly, 24 for daily}

\item{plotTitle}{Title of the plot}

\item{dirres}{the degree resolution of the wind rose}
}
\value{
plotly wind rose
}
\description{
Create Wind Rose Plot
}
